feat: silex.sile no longer overrides SILE by default
The re-implementation of some classes, packages, etc. no longer
takes over that of SILE.
Users have to enforce the use of our replacements explicitly:
- require("silex") to enforce all of sile·x
- require("silex.override") to only enforce the replacements

The resilient.sile module already does the first, so gets all of
the sile·x features when the resilient book class is used, but
other classes and uses are no longer affected without the user
explicit will.